requirements.txt
aiogram
python-dotenv
google-generativeai
PyPDF2
python-docx